TextInput: Remove PropTypes, NativeMethodsMixin; Convert to ES6 class
